+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+
+- The operator \= is now mapped to the built-in neq operator implementation.
+
 ### Security
 
 ## [1.1.0] - 2021-04-24